it can calculate now :)

+++ sukacheva.maria/S2/DataProcessing.hpp
@@ -9,6 +9,6 @@ namespace sukacheva {
   Queue< std::string > makePostfix(Queue< std::string >& infix);
   bool isBinaryOperations(std::string example);
   bool isNumber(std::string example);
-  size_t calculate(Queue< std::string >& postfix);
+  long long calculate(Queue< std::string >& postfix);
 }
 #endif
